local-model-rag/requirements.txt
ollama
python-dotenv
pandas
numpy
faiss-cpu
